refactor(core/form/inputs): deprecate block marker callbacks and support
