+++ packages/sanity/src/core/form/inputs/PortableText/text/TextBlock.tsx
@@ -20,7 +20,7 @@ import {
 import {useFormBuilder} from '../../../useFormBuilder'
 import {BlockActions} from '../BlockActions'
 import {ReviewChangesHighlightBlock, StyledChangeIndicatorWithProvidedFullPath} from '../_common'
-import {RenderBlockActionsCallback} from '../types'
+import {RenderBlockActionsCallback} from '../../../types/_transitional'
 import {useMemberValidation} from '../hooks/useMemberValidation'
 import {usePortableTextMarkers} from '../hooks/usePortableTextMarkers'
 import {usePortableTextMemberItem} from '../hooks/usePortableTextMembers'
