+++ packages/sanity/src/core/form/inputs/index.ts
@@ -1,5 +1,5 @@
 export type {PortableTextEditorElement} from './PortableText/Compositor'
-export type {RenderBlockActionsCallback, RenderBlockActionsProps} from './PortableText/types'
+export type {RenderBlockActionsCallback, RenderBlockActionsProps} from '../types/_transitional'
 export * from './files/types'
 export * from './PortableText/PortableTextInput'
 export {PortableTextInput as BlockEditor} from './PortableText/PortableTextInput'
