+++ packages/sanity/src/core/form/types/_transitional.ts
@@ -4,6 +4,7 @@ import {
   ObjectField,
   ObjectSchemaType,
   Path,
+  PortableTextBlock,
 } from '@sanity/types'
 import React from 'react'
 import {PatchEvent} from '../patch'
@@ -15,18 +16,41 @@ import {ItemProps} from './itemProps'
 /**
  * Export `PortableTextMarker` so it can be used to build custom Portable Text markers.
  *
- *
- * @hidden
  * @beta
+ * @hidden
+ * @deprecated - use `renderBlock`, `renderInlineBlock`, `renderAnnotation` interfaces instead
  */
 export type RenderCustomMarkers = (markers: PortableTextMarker[]) => React.ReactNode
 
 /**
- * A generic marker for attaching metadata to specific nodes of the Portable Text input.
- *
  *
  * @hidden
  * @beta
+ */
+export interface RenderBlockActionsProps {
+  block: PortableTextBlock
+  value: PortableTextBlock[] | undefined
+  set: (block: PortableTextBlock) => void
+  unset: () => void
+  insert: (block: PortableTextBlock | PortableTextBlock[]) => void
+}
+
+/**
+ *
+ * @hidden
+ * @beta
+ */
+export type RenderBlockActionsCallback = (props: RenderBlockActionsProps) => React.ReactNode
+
+/**
+ * A generic marker for attaching metadata to specific nodes of the Portable Text input.
+ *
+ * @beta
+ * @hidden
+ * @deprecated - use `renderBlock`, `renderInlineBlock`, `renderAnnotation` interfaces instead
+ * @param type - a type name for this marker
+ * @param data - some data connected to this marker
+ * @param path - the path to the Portable Text content connected to this marker
  */
 export interface PortableTextMarker {
   type: string
